version 1.0

+++ src/main/java/tn/esprit/spring/repository/EntrepriseRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import tn.esprit.spring.entities.Entreprise;
 
 @Repository
-public interface EntrepriseRepository extends  CrudRepository<Entreprise, Long> {
+public interface EntrepriseRepository extends  CrudRepository<Entreprise, Integer> {
 
 	
 }
